LIME-4: * fixed button types
