+++ frontend/src/bundles/auth/components/sign-in-form/sign-in-form.tsx
@@ -1,5 +1,8 @@
-import { Button } from '~/bundles/common/components/components.js';
-import { ButtonSize, ButtonVariant } from '~/bundles/common/types/types.js';
+import {
+    Button,
+    ButtonSize,
+    ButtonVariant,
+} from '~/bundles/common/components/components.js';
 
 type Properties = {
     onSubmit: () => void;
@@ -11,8 +14,8 @@ const SignInForm: React.FC<Properties> = () => (
         <form>
             <Button
                 label="Sign in"
-                variant={ButtonVariant.Primary}
-                size={ButtonSize.Medium}
+                variant={ButtonVariant.PRIMARY}
+                size={ButtonSize.MEDIUM}
             />
         </form>
     </>
