+++ frontend/src/bundles/common/components/components.ts
@@ -1,4 +1,4 @@
-export { Button } from './button/button.js';
+export { Button, ButtonSize, ButtonVariant } from './button/button.js';
 export { Input } from './input/input.js';
 export { Link } from './link/link.js';
 export { RouterProvider } from './router-provider/router-provider.js';
