+++ frontend/src/bundles/common/helpers/get-valid-class-names/get-valid-class-names.ts
