+++ frontend/src/bundles/common/helpers/helpers.ts
@@ -1,2 +1,2 @@
-export { getValidClassNames } from './create-class-names/get-valid-class-names.js';
+export { getValidClassNames } from './get-valid-class-names/get-valid-class-names.js';
 export { configureString } from 'shared';
